moved function back to their dashboard files and fixed button area issue
